Adjust documentation items
I think we can make do with one "update docs" item which can be checked as soon as the changes are understood and taken into account. That is, for new content, the new docs page does not need to be published before a PR can be merged.

If the docs content is added later (which I think will more often be the case now that we have a separate docs team), I'll take over the responsibility to add links and update the examples README once the content is published.

What do you think?

Signed-off-by: Sabine <sabine.nyholm@neptune.ai>